+++ codex_swap.py
@@ -29,7 +29,7 @@ from xswap_upgrade import UpgradeError, upgrade
 from xswap_alert import AlertError
 from xswap_alert import install as alert_install, status as alert_status, uninstall as alert_uninstall
 
-__version__ = "0.7.3"
+__version__ = "0.7.4"
 
 
 class SwapError(Exception):